apps/whatsapp-agent/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
gunicorn>=23.0
twilio>=9.3
openai>=1.55
httpx>=0.27
pydantic>=2.9
pydantic-settings>=2.6
python-dotenv>=1.0
python-multipart>=0.0.12
asyncpg>=0.30
psycopg[binary,pool]>=3.2
supabase>=2.10
pypdf>=5.1
pdf2image>=1.17
pillow>=11.0
cryptography>=43.0
structlog>=24.4
sentry-sdk[fastapi]>=2.18
posthog>=3.7
tenacity>=9.0
redis>=5.2
